requirements.txt
kafka-python>=2.0.0
pyspark>=3.4.0
